+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.13.0 February 15, 2026
 - Added
     - Tab shows player list
+    - Added chatbox to not pause game
 - Changed
     - Chat now only stores 50 most recent messages